Rename the package to kobuki_velocity_smoother.
This is to disambiguate it from other velocity smoother
packages that exist in the ROS ecosystem.
